+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Migrated to hyper. Works as expected.
 * Dependencies cleared.
 * Logs cleared and switched to Compact log mode.
+* Nested json structure support.
 #### v0.0.1-alpha.1
 * added some information to **Cargo.toml** file.
 * added example main.rs test.